Separate time zone related code into Timezone.cpp
