+++ FlapFunctions.h
@@ -16,13 +16,8 @@ int getSuggestedOffset(int letterIndex);
 int translateLetterToIndex(char letterchar);
 char translateIndextoLetter(int index);
 void showNewData(String message);
-String getDateString();
-String getClockString();
-void showDate();
 void setOfflineClock(char *clock);
 void showOfflineClock();
-void showClock();
-void updateTimezone();
 void showMessage(String message, int flapRpm);
 UnitState fetchUnitState(int unitAddr);
 UnitState *getUnitStates();
