+++ Timezone.h
@@ -1,0 +1,10 @@
+#ifndef TIMEZONE_H
+#define TIMEZONE_H
+
+#include <Arduino.h>
+
+String getDateString();
+String getClockString();
+void applyUserTimezone();
+
+#endif // TIMEZONE_H
